pyproject.toml
[project]
name = "ddd-internet-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "Yauheni Mukhin",email = "evgen3email@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "pydantic[email] (>=2.10.4,<3.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
